+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Devop
+
 - Add translation guidelines to dapp_integration.md and PR guide [#1617](https://github.com/MyEtherWallet/MyEtherWallet/pull/1617)
+- Add rel="noopener noreferrer" to links[#1618](https://github.com/MyEtherWallet/MyEtherWallet/pull/1618)
 
 ### Release v5.2.2